+++ oras/oci.py
@@ -11,6 +11,7 @@ import jsonschema
 import oras.defaults
 import oras.schemas
 import oras.utils
+from oras.logger import logger
 
 EmptyManifest = {
     "schemaVersion": 2,
@@ -30,9 +31,19 @@ class Annotations:
         self.lookup = {}
         self.load(filename)
 
+    def add(self, section, key, value):
+        """
+        Add key/value pairs to a named section.
+        """
+        if section not in self.lookup:
+            self.lookup[section] = {}
+        self.lookup[section][key] = value
+
     def load(self, filename: str):
         if filename and os.path.exists(filename):
             self.lookup = oras.utils.read_json(filename)
+        if filename and not os.path.exists(filename):
+            logger.exit(f"Annotation file {filename} does not exist.")
 
     def get_annotations(self, section: str) -> dict:
         """